Move story data
