+++ src/components/Profile/Details/Details.tsx
@@ -1,11 +1,11 @@
 import React from "react";
-import { devStory } from "./data/dev-story";
+import { devStory } from "../../../data/dev-story";
 import ExperienceItem from "./ExperienceItem";
 import ProjectItem from "./ProjectItem";
 
 export default function Header() {
   return (
-    <main className="container mx-auto mt-8 md:flex md:px-4 md:pb-4 max-w-4xl">
+    <main className="mt-8 md:flex md:px-4 md:pb-4">
       <div className="mx-4 flex-1">
         <SectionTitle title="Projects" />
         <div>
