+++ src/components/Profile/Profile.tsx
@@ -3,9 +3,9 @@ import Header from "./Header";
 
 export default function Profile() {
   return (
-    <>
+    <div className="container mx-auto lg:max-w-4xl">
       <Header />
       <Details />
-    </>
+    </div>
   );
 }
