+++ src/components/common/Tags/Tags.tsx
@@ -1,10 +1,11 @@
 type TagsProps = {
   tags: string[];
+  className?: string;
 };
 
 export default function Tags(props: TagsProps) {
   return (
-    <div className="flex flex-wrap">
+    <div className={`flex flex-wrap ${props.className}`}>
       {props.tags.map((tag) => (
         <TagItem title={tag} />
       ))}
@@ -12,9 +13,9 @@ export default function Tags(props: TagsProps) {
   );
 }
 
-function TagItem({ title }: { title: string} ) {
+function TagItem({ title }: { title: string }) {
   return (
-    <span className="text-xs bg-cyan-500 text-white px-1 py-0.5 mr-1 mt-1 rounded-sm">
+    <span className="text-xs bg-slate-50 text-slate-600 px-1 py-0.5 mr-1 mt-1 rounded-sm">
       {title}
     </span>
   );
