+++ src/data/dev-story.ts
@@ -1,4 +1,26 @@
-import { DevStory } from "./dev-story.d";
+export type DevStory = {
+  projects: Project[];
+  experiences: Experience[];
+};
+
+export type Project = {
+  year: number;
+  month: number;
+  what: string;
+  about: string;
+  tags: string[];
+  link: string | null;
+};
+
+export type Experience = {
+  yearStart: number;
+  monthStart: number;
+  yearEnd: number | null;
+  monthEnd: number | null;
+  where: string;
+  what: string;
+  tags: string[];
+};
 
 export const devStory: DevStory = {
   projects: [
@@ -262,7 +284,7 @@ export const devStory: DevStory = {
       yearEnd: 2011,
       monthEnd: 4,
       where: "Pizza Hut",
-      what: "Pizza Delivery Boy",
+      what: "Pizza Delivery Boy 🍕",
       tags: [],
     },
     {
@@ -271,7 +293,7 @@ export const devStory: DevStory = {
       yearEnd: 2009,
       monthEnd: 10,
       where: "West Midlands Safari Park",
-      what: "Theme Park Ride Operator",
+      what: "Theme Park Ride Operator 🎢",
       tags: [],
     },
   ],
